major backend refactoring, uses smaller modules, consistent names style cache loads cached code from disk on initialization better metadata cache reusage
