+++ packages/backend/src/controllers/index.ts
@@ -1,4 +1,10 @@
-export { generic as renderGeneric, renderSplice } from './renderers';
-export { details as styleDetails, all as allStyles } from './style';
-export { spliceMetadata } from './metadata';
 export { allowlist } from './allowlist';
+export { allStyles } from './allStyles';
+export { extractColors } from './extractColors';
+export { nftMetadata } from './nftMetadata';
+export { proxy } from './proxy';
+export { renderGeneric } from './renderGeneric';
+export { renderSplice } from './renderSplice';
+export { spliceMetadata } from './spliceMetadata';
+export { styleDetails } from './styleDetails';
+
