+++ packages/backend/src/lib/Artwork.ts
@@ -1,6 +1,6 @@
-import { BANNER_DIMS, DrawArgs, NFTItem, NFTMetaData } from '@splicenft/common';
-import { fetchOriginMetadata } from '../controllers/metadata';
+import { BANNER_DIMS, DrawArgs, NFTItem } from '@splicenft/common';
 import * as Cache from './Cache';
+import { fetchOriginMetadata } from './fetchOriginMetadata';
 import { IImageCallback } from './ImageCallback';
 import Metadata from './Metadata';
 import { RenderAndCache } from './render';
@@ -9,40 +9,37 @@ import { StyleMetadataCache } from './StyleCache';
 export default async function Artwork(
   styleCache: StyleMetadataCache,
   spliceTokenId: string,
-  callback: IImageCallback
+  callback: IImageCallback,
+  invalidate = false
 ): Promise<void> {
+  const imageKey = `${styleCache.network}/splice/images/${spliceTokenId}.png`;
+  if (invalidate) {
+    await Cache.remove(imageKey);
+  }
+
   const spliceMetadata = await Cache.withCache(
     `${styleCache.network}/splice/metadata/${spliceTokenId}.json`,
-    async () => await Metadata(styleCache, spliceTokenId)
+    async () => await Metadata(styleCache, spliceTokenId),
+    invalidate
   );
-  const key = `${styleCache.network}/splice/images/${spliceTokenId}.png`;
 
-  const stream = await Cache.lookupBinary(key);
+  const stream = await Cache.lookupBinary(imageKey);
   if (stream) {
     return callback(null, stream, []);
   }
 
-  const style = styleCache.getStyle(spliceMetadata.splice.style_token_id);
-  if (!style) throw new Error('no style on metadata ?!');
-
-  const renderer = await style.getRenderer();
-
   const firstOrigin = spliceMetadata.splice.origins[0];
-  const cachedOriginMetadata = await Cache.withCache<NFTMetaData>(
-    `${styleCache.network}/nft/${firstOrigin.collection}/${firstOrigin.token_id}/metadata.json`,
-    () =>
-      fetchOriginMetadata(
-        styleCache.network,
-        firstOrigin.collection,
-        firstOrigin.token_id
-      ),
-    false
+  const originMetadata = await fetchOriginMetadata(
+    styleCache.network,
+    firstOrigin.collection,
+    firstOrigin.token_id,
+    invalidate
   );
 
   const nftItem: NFTItem = {
     contract_address: firstOrigin.collection,
     token_id: firstOrigin.token_id.toString(),
-    metadata: cachedOriginMetadata
+    metadata: originMetadata
   };
   const drawArgs: DrawArgs = {
     dim: BANNER_DIMS,
@@ -52,5 +49,10 @@ export default async function Artwork(
       nftItem
     }
   };
-  RenderAndCache(key, drawArgs, renderer, callback);
+
+  const style = styleCache.getStyle(spliceMetadata.splice.style_token_id);
+  if (!style) throw new Error('no style on metadata ?!');
+  const renderer = await style.getRenderer();
+
+  RenderAndCache(imageKey, drawArgs, renderer, callback);
 }
