+++ packages/backend/src/lib/Metadata.ts
@@ -8,14 +8,16 @@ import {
 import { BigNumber } from 'ethers';
 import { Readable } from 'stream';
 import * as Cache from './Cache';
-import { extractOriginFeatures, getOriginMetadata } from './Origin';
+import { extractOriginFeatures } from './extractOriginFeatures';
+import { fetchOriginMetadata } from './fetchOriginMetadata';
 import { Render } from './render';
 import { getSplice } from './SpliceContracts';
 import { StyleMetadataCache } from './StyleCache';
 
 const Metadata = async (
   styleCache: StyleMetadataCache,
-  spliceTokenId: string
+  spliceTokenId: string,
+  invalidate = false
 ): Promise<SpliceNFT> => {
   const splice = await getSplice(styleCache.network);
   const provenance = await splice.getProvenance(BigNumber.from(spliceTokenId));
@@ -23,26 +25,24 @@ const Metadata = async (
     throw new Error(`no provenance for token ${spliceTokenId}`);
   }
 
-  const style = styleCache.getStyle(provenance.style_token_id);
-  if (!style) throw new Error(`style token seems corrupt`);
-
   const firstOrigin = provenance.origins[0];
   if (!firstOrigin) throw new Error(`no origin for splice`);
 
-  const originNftContract = splice.getOriginNftContract(firstOrigin.collection);
-  const originNftName = await originNftContract.name();
-
-  const originMetadata = await getOriginMetadata(
-    originNftContract,
+  const originMetadata = await fetchOriginMetadata(
+    styleCache.network,
+    firstOrigin.collection,
     firstOrigin.token_id
   );
 
-  if (!originMetadata) throw new Error(`couldnt get origin metadata`);
-
   const originFeatures = await Cache.withCache<Transfer.OriginFeatures>(
     `${styleCache.network}/nft/${firstOrigin.collection}/${firstOrigin.token_id}/features.json`,
-    async () => extractOriginFeatures(firstOrigin, originMetadata)
+    async () => extractOriginFeatures(firstOrigin, originMetadata),
+    invalidate
   );
+
+  const originNftName = await splice
+    .getOriginNftContract(firstOrigin.collection)
+    .name();
 
   const nftItem: NFTItem = {
     contract_address: firstOrigin.collection,
@@ -51,7 +51,10 @@ const Metadata = async (
   };
 
   //we're invoking the renderer here to get the traits out of the rendered style.
+  const style = styleCache.getStyle(provenance.style_token_id);
+  if (!style) throw new Error(`style token seems corrupt`);
   const renderer = await style.getRenderer();
+
   return new Promise((resolve, reject) => {
     Render(
       {
